Move trait map into a submodule.
