+++ sway-core/src/semantic_analysis/namespace/trait_map/map.rs
