+++ sway-core/src/semantic_analysis/namespace/trait_map/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod map;
